Build one single shared library for the end-user
The goal is to provide just one library that needs to be linked with the
project using litesql, with a simple -llitesql. Each backend-specific
library is built statically and used only to build the final litesql
library.

An option is available, so the user can choose to build the main litesql
library dynamically or statically.